fix: query files can use streaming as long as they don't have named parameters

+++ test/helpers/scripts/functions/psParams.sql
@@ -1,1 +1,1 @@
-SELECT $1 + $2;
+SELECT $1::INT + $2::INT;
